+++ src/dyninst/test_thread_6_mutatee.c
@@ -64,7 +64,7 @@ int test_thread_6_mutatee() {
    initThreads();
 
    for (i=0; i<NTHRD; i++)  {
-       threads[i] = spawnNewThread((void *) init_func, (void *) i);
+      threads[i] = spawnNewThread((void *) init_func, (void *) (long) i);
    }
 
    while (!startedall) {
